second

+++ src/main/java/com/project/myproject/dto/UserAccountDto.java
@@ -3,14 +3,15 @@ package com.project.myproject.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor
 @Getter
 public class UserAccountDto extends UserDto {
-    private double balance;
+    private BigDecimal balance;
 
-    public UserAccountDto(String name, double exp, String role, UserImageDto image, double balance) {
+    public UserAccountDto(String name, double exp, String role, UserImageDto image, BigDecimal balance) {
         super(name, exp, role, image);
         this.balance = balance;
     }
